pyproject.toml
[tool.poetry]
name = "cafs-backend"
version = "0.1.0"
description = ""
authors = ["Phagatanga"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
djangorestframework-simplejwt = "^5.2.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
